Crate hacspec

hacspec consists of three parts:

hacspec library

syntax checker

compiler

The hacspec library

The hacspec library implements a comprehensive set of functions to implement succinct cryptographic specs that can be compiled to formal languages such as F*.

The syntax checker

TODO:

describe clippy extension

add cargo hacspec check command

The compiler

TODO:

define compiler interface

add cargo hacspec fstar command

Modules

prelude
This module conveniently exports common subroutines necessary for hacspecs

Macros

bytes

hacspec_crates	Deprecated

hacspec_imports	Deprecated

random_bytes

vlbytes

Structs

ByteSlice

Bytes
Variable length byte arrays.

Traits

ByteArray
Common trait for all byte arrays.

Functions

from_u32l

get_random_bytes

hex_string_to_bytes

to_array

to_u32l