Module hacspec::array

Arrays

This module implements fixed-length arrays and utility functions for it.

Note that implementations have to be created with one of the provided macros such that there's no documentation in here.

You can find examples for the different types of arrays here:

DocSecretBytes for bytes!(DocSecretBytes, 64)

DocPublicBytes for public_bytes!(DocPublicBytes, 64)

DocSecretArray for array!(DocSecretArray, 64, U32)

DocPublicArray for array!(DocPublicArray, 64, u32)

Note that all macros starting with an underscore (_array_base etc.) are note intended for public use. Unfortunately it's not possible to hide them.