Module hacspec::math_util

Functions

cadd
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

cmul
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

cset_bit
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

csub
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

cswap
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

cswap_bit
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

ct_div
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

extended_euclid
Constant-time extended euclidean algorithm to compute the inverse of x in yℤ[x] x.len() and degree of y are assumed to be public See recipx in Figure 6.1 of https://eprint.iacr.org/2019/266

poly_add

poly_div
Euclidean algorithm to compute quotient q and remainder r of x/y. The length of x and degree of y are assumed to be public

poly_mul
Polynomial multiplication using sparse multiplication. This can be more efficient than operand scanning but also prone to side-channel attacks.

poly_sub