Module hacspec::poly

Polynomials

This module implements polynomials ℤn[x]/mℤ[x]. Polynomials are variable sized only for now.

Coefficients are currently stored as u128 or i128.

TODO: If necessary, we could extend the definition to larger integers.

Three different types of polynomials are supported:

Polynomial rings over ℤ

Quotient ring over ℤn

Quotient ring over ℤn/mℤ

Polynomial rings over ℤ

This most basic form is implemented over basic sequences Seq<T> and arrays array!(...). Addition, Subtraction, Multiplication, and Division with remainder are supported.

Note: This is currently only implemented for Seq<u128> and Seq<i128>.

Examples

use hacspec::prelude::*;

let x = Seq::<u128>::from_array(&[5, 2, 7, 8, 9]); // 5 + 2x + 7x² + 8x³ + 9x⁴
let y = Seq::<u128>::from_array(&[2, 1, 0, 2, 4]); // 2 + 1x       + 2x³ + 4x⁴
let z = x.clone() * y.clone();
let z = x.clone() + y.clone();
let z = x.clone() - y.clone();
// You can not divide polynomials over ℤ. This will panic.
// let (q, r) = x.clone() / y.clone();

Quotient ring over ℤn

Quotient ring arithmetic is implemented on a separate type Poly<T>, which behaves similar to sequences but can store the modulus n as well. Addition, Subtraction, Multiplication, and Division with remainder are supported.

Note: This is currently only implemented for Poly<u128> and Poly<i128>.

Examples

use hacspec::prelude::*;

let x = Poly::<u128>::from_array(&[5, 2, 7, 8, 9], 11); // 5 + 2x + 7x² + 8x³ + 9x⁴
let y = Poly::<u128>::from_array(&[2, 1, 0, 2, 4], 11); // 2 + 1x       + 2x³ + 4x⁴
let z = x.clone() * y.clone();
let z = x.clone() + y.clone();
let z = x.clone() - y.clone();
let (q, r) = x.clone() / y.clone();

Quotient ring over ℤn/mℤ[x]

Quotient ring arithmetic is implemented on a separate type Poly<T>, which behaves similar to sequences but can store the modulus n and irreducible polynomial irr as well. Addition, Subtraction, Multiplication, and Division with remainder are supported.

Note: This is currently only implemented for Poly<u128> and Poly<i128>.

Examples

use hacspec::prelude::*;

let irr = [1, 3, 5, 0, 8, 6];
let x = Poly::<u128>::new_full(&irr, &[5, 2, 7, 8, 9], 11); // 5 + 2x + 7x² + 8x³ + 9x⁴
let y = Poly::<u128>::new_full(&irr, &[2, 1, 0, 2, 4], 11); // 2 + 1x       + 2x³ + 4x⁴
let z = x.clone() * y.clone();
let z = x.clone() + y.clone();
let z = x.clone() - y.clone();
let (q, r) = x.clone() / y.clone();

Structs

Poly
The poly struct. Note that while this is a polynomial over ℤn[x]/mℤ[x] it is not necessarily reduced by mℤ[x], i.e. poly might be larger. Call reduce to make sure poly is in ℤn/mℤ.

Traits

Integer
Trait that needs to be implemented by all integers that are used as coefficients. This is done here for ℤn over i128 and u128.

TRestrictions
Traits that have to be implemented by the type used for coefficients.

Functions

random_poly