Struct hacspec::poly::Poly

pub struct Poly<T: TRestrictions<T>> { /* fields omitted */ }

The poly struct. Note that while this is a polynomial over ℤn[x]/mℤ[x] it is not necessarily reduced by mℤ[x], i.e. poly might be larger. Call reduce to make sure poly is in ℤn/mℤ.

Methods

impl<T: TRestrictions<T>> Poly<T>

pub fn from_array(p: &[u128], n: u128) -> Self

pub fn new_full(irr_in: &[u128], p: &[u128], n_in: u128) -> Self

pub fn new(irr_in: &[u128], p: &[u128]) -> Self

pub fn new_signed(irr_in: &[i128], p: &[i128]) -> Self

pub fn new_signed_full(irr_in: &[i128], p: &[i128], n_in: i128) -> Self

pub fn new_monomial(irr_in: &[u128], c: T, d: usize) -> Self

pub fn new_poly(self, p: &[u128]) -> Self

pub fn random(irr_in: Seq<T>, r: Range<i128>, n_in: T) -> Self

Generate random polynomial with given coefficient bounds and irreducible. Note that this requires min and max to be i128. For random coefficients in T that don't fit in i128 use generators on T.

pub fn leading_coefficient(self) -> (usize, T)

Returns the leading coefficient of this polynomial and it's index.

pub fn reduce(self) -> Self

Reduce self, i.e. self.poly by the irreducible. Returns self.poly % self.irr.

pub fn pad(self, l: usize) -> Self

Pad self.poly to length l with zeroes.

pub fn truncate(&self) -> Self

Truncate self.poly, removing trailing zeroes.

pub fn is_zero(self) -> bool

Returns true if this is an all-zero polynomial.

pub fn euclid_div(self, rhs: Poly<T>) -> (Poly<T>, Poly<T>)

Euclidean division returning (q, r)

pub fn inv(self) -> Self

Invert this polynomial.