Trait hacspec::poly::Integer

pub trait Integer<T> {
    fn from_literal(x: u128) -> T;

    fn from_signed_literal(x: i128) -> T;

    fn inv(x: T, n: T) -> T;

    fn max() -> T;

    fn sub_lift(self, rhs: T, n: T) -> T;

    fn sub_mod(self, rhs: T, n: T) -> T;

    fn add_mod(self, rhs: T, n: T) -> T;

    fn mul_mod(self, rhs: T, n: T) -> T;

    fn rem(self, n: T) -> T;

    fn abs(self) -> T;
}

Trait that needs to be implemented by all integers that are used as coefficients. This is done here for ℤn over i128 and u128.

Required methods

fn from_literal(x: u128) -> T

fn from_signed_literal(x: i128) -> T

fn inv(x: T, n: T) -> T

fn max() -> T

fn sub_lift(self, rhs: T, n: T) -> T

Lift the possibly negative result back up mod n.

fn sub_mod(self, rhs: T, n: T) -> T

Compute (self - rhs) % n.

fn add_mod(self, rhs: T, n: T) -> T

(self + rhs) % n

fn mul_mod(self, rhs: T, n: T) -> T

(self * rhs) % n

fn rem(self, n: T) -> T

self % n

fn abs(self) -> T

Loading content...