Module hacspec::prelude

This module conveniently exports common subroutines necessary for hacspecs

use hacspec::prelude::*;

Re-exports

pub use crate::array::*;

pub use crate::poly::*;

pub use crate::seq::*;

pub use crate::util::*;

pub use crate::machine_integers::*;

pub use crate::math_integers::*;

pub use crate::math_util::*;

pub use crate::traits::*;

pub use crate::vec_integers::*;

pub use crate::transmute::*;

pub use crate::*;

pub use num;

pub use rand;

Modules

abstract_int

fmt
Utilities for formatting and printing Strings.

nat_mod

traits

Macros

Debug
Derive macro generating an impl of the trait Debug.

PartialEq
Derive macro generating an impl of the trait PartialEq.

abstract_int

abstract_nat_mod

abstract_public

abstract_public_modular_integer

abstract_public_nat_mod

abstract_secret

abstract_secret_modular_integer

abstract_signed

abstract_signed_public_integer

abstract_signed_secret_integer

abstract_unsigned

abstract_unsigned_public_integer

abstract_unsigned_secret_integer

define_abstract_integer_checked
Defines a bounded natural integer with regular arithmetic operations, checked for overflow and underflow.

define_refined_modular_integer
Defines a bounded natural integer with modular arithmetic operations

modular_integer

Structs

BigInt
A big signed integer type.

BigUint
A big unsigned integer type.

I8

I16

I32

I64

I128

ParseIntError
An error which can be returned when parsing an integer.

Range
A (half-open) range bounded inclusively below and exclusively above (start..end).

RangeFrom
A range only bounded inclusively below (start..).

RangeFull
An unbounded range (..).

RangeInclusive
A range bounded inclusively below and above (start..=end).

RangeTo
A range only bounded exclusively above (..end).

RangeToInclusive
A range only bounded inclusively above (..=end).

U8

U16

U32

U64

U128

Enums

Bound
An endpoint of a range of keys.

Ordering
An Ordering is the result of a comparison between two values.

GeneratorState	Experimental
The result of a generator resumption.

Sign
A Sign is a BigInt's composing element.

Traits

Add
The addition operator +.

AddAssign
The addition assignment operator +=.

BitAnd
The bitwise AND operator &.

BitAndAssign
The bitwise AND assignment operator &=.

BitOr
The bitwise OR operator |.

BitOrAssign
The bitwise OR assignment operator |=.

BitXor
The bitwise XOR operator ^.

BitXorAssign
The bitwise XOR assignment operator ^=.

CheckedSub
Performs subtraction that returns None instead of wrapping around on underflow.

Debug
? formatting.

Deref
Used for immutable dereferencing operations, like *v.

DerefMut
Used for mutable dereferencing operations, like in *v = 1;.

Div
The division operator /.

DivAssign
The division assignment operator /=.

Drop
Custom code within the destructor.

Fn
The version of the call operator that takes an immutable receiver.

FnMut
The version of the call operator that takes a mutable receiver.

FnOnce
The version of the call operator that takes a by-value receiver.

Index
Used for indexing operations (container[index]) in immutable contexts.

IndexMut
Used for indexing operations (container[index]) in mutable contexts.

Mul
The multiplication operator *.

MulAssign
The multiplication assignment operator *=.

Neg
The unary negation operator -.

Not
The unary logical negation operator !.

Num
The base trait for numeric types, covering 0 and 1 values, comparisons, basic numeric operations, and string conversion.

One
Defines a multiplicative identity element for Self.

PartialEq
Trait for equality comparisons which are partial equivalence relations.

RangeBounds
RangeBounds is implemented by Rust's built-in range types, produced by range syntax like .., a.., ..b, ..=c, d..e, or f..=g.

Rem
The remainder operator %.

RemAssign
The remainder assignment operator %=.

Shl
The left shift operator <<. Note that because this trait is implemented for all integer types with multiple right-hand-side types, Rust's type checker has special handling for _ << _, setting the result type for integer operations to the type of the left-hand-side operand. This means that though a << b and a.shl(b) are one and the same from an evaluation standpoint, they are different when it comes to type inference.

ShlAssign
The left shift assignment operator <<=.

Shr
The right shift operator >>. Note that because this trait is implemented for all integer types with multiple right-hand-side types, Rust's type checker has special handling for _ >> _, setting the result type for integer operations to the type of the left-hand-side operand. This means that though a >> b and a.shr(b) are one and the same from an evaluation standpoint, they are different when it comes to type inference.

ShrAssign
The right shift assignment operator >>=.

Sub
The subtraction operator -.

SubAssign
The subtraction assignment operator -=.

CoerceUnsized	Experimental
Trait that indicates that this is a pointer or a wrapper for one, where unsizing can be performed on the pointee.

DispatchFromDyn	Experimental
This is used for object safety, to check that a method's receiver type can be dispatched on.

Generator	Experimental
The trait implemented by builtin generator types.

Rng
An automatically-implemented extension trait on [RngCore] providing high-level generic methods for sampling values and other convenience methods.

Sub
The subtraction operator -.

Try	Experimental
A trait for customizing the behavior of the ? operator.

Zero
Defines an additive identity element for Self.

Functions

min
Compares and returns the minimum of two values.

one
Returns the multiplicative identity, 1.

zero
Returns the additive identity, 0.

Attribute Macros

external
Function that is not part of the language but is offered as a helper for tests, etc.

internal
Internal function used in primitive implementations but that is not part of the language.

library
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

primitive
Primitive function, part of the formalization of the language.

to_remove
Function that should be removed from the library.