Struct hacspec::seq::Seq

pub struct Seq<T: Copy> { /* fields omitted */ }

Variable length byte arrays.

Methods

impl<T: Copy + Default> Seq<T>

pub fn new(l: usize) -> Self

Get a new sequence of capacity l.

pub fn from_array(v: &[T]) -> Self

Get a new sequence from array v.

pub fn len(&self) -> usize

Get the size of this sequence.

pub fn update<A: SeqTrait<T>>(self, start: usize, v: A) -> Self

Update this sequence with v starting at start.

Examples

use hacspec::prelude::*;

let mut s = Seq::<u8>::new(5);
let tmp = Seq::<u8>::from_array(&[2, 3]);
s = s.update(2, tmp);
assert_eq!(s, Seq::<u8>::from_array(&[0, 0, 2, 3, 0]));

pub fn update_sub<A: SeqTrait<T>>(
self,
start_out: usize,
v: A,
start_in: usize,
len: usize
) -> Self

Update this sequence with l elements of v, starting at start_in, at start_out.

Examples

use hacspec::prelude::*;

let mut s = Seq::<u8>::new(5);
let tmp = Seq::<u8>::from_array(&[2, 3]);
s = s.update_sub(2, tmp, 1, 1);
assert_eq!(s, Seq::<u8>::from_array(&[0, 0, 3, 0, 0]));

pub fn update_element(self, start_out: usize, v: T) -> Self

Update this sequence with v at position start_out.

Examples

use hacspec::prelude::*;

let mut s = Seq::<u8>::new(5);
s = s.update_element(4, 7);
assert_eq!(s, Seq::<u8>::from_array(&[0, 0, 0, 0, 7]));

pub fn set_index(self, i: usize) -> Self

Reset the sequence index.

Examples

use hacspec::prelude::*;

let mut s = Seq::<u8>::new(5);
s = s.set_index(3);
s = s.push(Seq::<u8>::from_array(&[4, 5]));
assert_eq!(s, Seq::<u8>::from_array(&[0, 0, 0, 4, 5]));

pub fn push<A: SeqTrait<T>>(self, v: A) -> Self

Push v to this sequence and move idx according to v.len().

Examples

use hacspec::prelude::*;

let mut s = Seq::<u8>::new(5);
s = s.set_index(3);
s = s.push(Seq::<u8>::from_array(&[4, 5]));
assert_eq!(s, Seq::<u8>::from_array(&[0, 0, 0, 4, 5]));

pub fn push_sub<A: SeqTrait<T>>(self, v: A, start: usize, l: usize) -> Self

Push l elements from v to this sequence and move idx according to l.

Examples

use hacspec::prelude::*;

let mut s = Seq::<u8>::new(5);
s = s.set_index(3);
s = s.push_sub(Seq::<u8>::from_array(&[4, 5]), 1, 1);
assert_eq!(s, Seq::<u8>::from_array(&[0, 0, 0, 5, 0]));

pub fn sub(self, start_out: usize, len: usize) -> Self

pub fn from_sub<A: SeqTrait<T>>(input: A, r: Range<usize>) -> Self

pub fn chunks<'a>(
&'a self,
chunk_size: usize
) -> impl Iterator<Item = (usize, Seq<T>)> + 'a

impl Seq<U8>

pub fn random(l: usize) -> Self

pub fn to_hex(&self) -> String

impl Seq<u8>

pub fn to_hex(&self) -> String