Struct hacspec::u64Word

pub struct u64Word(pub [u8; 8]);

Fixed length byte array. Because Rust requires fixed length arrays to have a known size at compile time there's no generic fixed length byte array here. Use this to define the fixed length byte arrays needed in your code.

Methods

impl u64Word

pub fn new() -> Self

pub fn capacity() -> usize

pub fn from_sub_pad<A: SeqTrait<u8>>(input: A, r: Range<usize>) -> Self

pub fn from_sub<A: SeqTrait<u8>>(input: A, r: Range<usize>) -> Self

pub fn copy_pad<A: SeqTrait<u8>>(v: A) -> Self

pub fn copy<A: SeqTrait<u8>>(v: A) -> Self

pub fn update<A: SeqTrait<u8>>(self, start: usize, v: A) -> Self

pub fn update_sub<A: SeqTrait<u8>>(
self,
start_out: usize,
v: A,
start_in: usize,
len: usize
) -> Self

pub fn len(&self) -> usize

pub fn to_bytes_be(&self) -> [u8; 8]

pub fn chunks<'a>(
&'a self,
chunk_size: usize
) -> impl Iterator<Item = (usize, Seq<u8>)> + 'a

Get an iterator over this array with chunks of size chunk_size.

Examples

use hacspec::prelude::*;

public_bytes!(Block, 5);
let a = Block::from([0, 1, 2, 3, 4]);
let mut a_chunks = a.chunks(2);
let a_chunk = a_chunks.next().unwrap();
assert_eq!(a_chunk.0, 2);
assert_eq!(a_chunk.1, Seq::<u8>::from_array(&[0, 1]));
let a_chunk = a_chunks.next().unwrap();
assert_eq!(a_chunk.0, 2);
assert_eq!(a_chunk.1, Seq::<u8>::from_array(&[2, 3]));
let a_chunk = a_chunks.next().unwrap();
assert_eq!(a_chunk.0, 1);
assert_eq!(a_chunk.1, Seq::<u8>::from_array(&[4]));

let a = Block::from([0, 1, 2, 3, 4]);
for (l, chunk) in a.chunks(2) {
    println!("{:x?}", chunk); // prints [0, 1], [2, 3], [4]
}

impl u64Word

pub fn random() -> u64Word

impl u64Word

pub fn get_random_vec(l: usize) -> Vec<u8>

impl u64Word

pub fn to_u32s_be(&self) -> [u32; 2]

pub fn to_hex(&self) -> String