Module hacspec::traits

This module specifies the Numeric trait.

Traits

Integer
This trait extends the Numeric trait and is implemented by all integer types. It offers bit manipulation, instantiation from literal, and convenient constants.

ModNumeric

Numeric
The Numeric trait has to be implemented by all numeric objects.

NumericCopy

PublicInteger

PublicIntegerCopy

SecretInteger

SecretIntegerCopy

SeqTrait
Common trait for all byte arrays and sequences.

SignedInteger

SignedIntegerCopy

UnsignedInteger

UnsignedIntegerCopy

UnsignedPublicInteger

UnsignedPublicIntegerCopy

UnsignedSecretInteger

UnsignedSecretIntegerCopy