Trait hacspec::traits::Integer

pub trait Integer: Numeric {
    const NUM_BITS: usize;

    fn ZERO() -> Self;

    fn ONE() -> Self;

    fn TWO() -> Self;

    fn from_literal(val: u128) -> Self;

    fn from_hex_string(s: &String) -> Self;

    fn get_bit(self, i: usize) -> Self { ... }

    fn set_bit(self, b: Self, i: usize) -> Self { ... }

    fn set(self, pos: usize, y: Self, yi: usize) -> Self { ... }

    fn rotate_left(self, n: usize) -> Self { ... }

    fn rotate_right(self, n: usize) -> Self { ... }
}

This trait extends the Numeric trait and is implemented by all integer types. It offers bit manipulation, instantiation from literal, and convenient constants.

Associated Constants

const NUM_BITS: usize

Loading content...
Required methods

fn ZERO() -> Self

fn ONE() -> Self

fn TWO() -> Self

fn from_literal(val: u128) -> Self

Get an integer with value val.

fn from_hex_string(s: &String) -> Self

Read a hex string (starting with 0x) into an Integer.

Loading content...
Provided methods

fn get_bit(self, i: usize) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

Get bit i of this integer.

fn set_bit(self, b: Self, i: usize) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

Set bit i of this integer to b and return the result. Bit b has to be 0 or 1.

fn set(self, pos: usize, y: Self, yi: usize) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

Set bit pos of this integer to bit yi of integer y.

fn rotate_left(self, n: usize) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

fn rotate_right(self, n: usize) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

Loading content...