Module hacspec::transmute

Functions

U16_to_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U16_to_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U16_from_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U16_from_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U32_to_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U32_to_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U32_from_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U32_from_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U64_to_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U64_to_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U64_from_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U64_from_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U128_to_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U128_to_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U128_from_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

U128_from_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u16_to_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u16_to_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u16_from_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u16_from_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u32_to_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u32_to_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u32_from_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u32_from_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u64_to_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u64_to_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u64_from_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u64_from_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u128_to_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u128_to_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u128_from_be_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

u128_from_le_bytes
Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.