Module hacspec::poly

Polynomials

This module implements polynomials ℤn[x]/mℤ[x]. Polynomials are variable sized only for now.

Coefficients are currently stored as u128 or i128.

TODO: If necessary, we could extend the definition to larger integers.

Three different types of polynomials are supported:

Polynomial rings over ℤ

Quotient ring over ℤn

Quotient ring over ℤn/mℤ

Polynomial rings over ℤ

This most basic form is implemented over basic sequences Seq<T>. Addition, Subtraction, Multiplication, and Division with remainder are supported.

Note: This is currently only implemented for Seq<u128> and Seq<i128>.

Functions

euclid_div
Euclidean algorithm to compute quotient q and remainder r of x/y.

extended_euclid
Extended euclidean algorithm to compute the inverse of x in yℤ[x]

leading_coefficient

poly_add

poly_mul
Polynomial multiplication using sparse multiplication. This can be more efficient than operand scanning but also prone to side-channel attacks.

poly_sub

random_poly