Trait hacspec::traits::Integer

pub trait Integer: Numeric {
    const NUM_BITS: u32;
    const ZERO: Self;
    const ONE: Self;
    const TWO: Self;

    fn from_literal(val: u128) -> Self;

    fn get_bit(self, i: u32) -> Self { ... }

    fn set_bit(self, b: Self, i: u32) -> Self { ... }

    fn set(self, pos: u32, y: Self, yi: u32) -> Self { ... }

    fn rotate_left(self, n: u32) -> Self { ... }

    fn rotate_right(self, n: u32) -> Self { ... }
}

This trait extends the Numeric trait and is implemented by all integer types. It offers bit manipulation, instantiation from literal, and convenient constants.

Associated Constants

const NUM_BITS: u32

const ZERO: Self

const ONE: Self

const TWO: Self

Loading content...
Required methods

fn from_literal(val: u128) -> Self

Get an integer with value val.

Loading content...
Provided methods

fn get_bit(self, i: u32) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

Get bit i of this integer.

fn set_bit(self, b: Self, i: u32) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

Set bit i of this integer to b and return the result. Bit b has to be 0 or 1.

fn set(self, pos: u32, y: Self, yi: u32) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

Set bit pos of this integer to bit yi of integer y.

fn rotate_left(self, n: u32) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

fn rotate_right(self, n: u32) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

Loading content...